Improved filtering of lines in .col files with checksums
I ran into a problem where, since I'd modified my .ggg files to
not write spt files, there was a line like:

./spt/ch4_5938/z 0

in the .col files that confused the existing check that a line split
on spaces had a length of 2 if it had a checksum + filename. I changed
it to use a regex to explicitly check that there is a 32-hex digit
checksum, followed one or more spaces, then a non-whitespace character
at the beginning of the line. This should be more robust against
unexpected line formats.