Added attribute to public files for data withheld
In the TCCON telecon on 2021-08-12, concern was raised over
what happens if a large amount of a sites' data is removed
by the release flagging. To abate this, we decided to add
an attribute to the public files that lists time periods
flagged out and points the user to the point of contact
for that site.

Also fixed the release flag attributes getting into the
public files.