Fix for prior index - use runlog to match up
With the previous version, there was a bad corner
case in the code that calculated the prior index.

In the Park Falls 2015 data, the .aia file had all
Si spectra and an instance (around spectrum
pa20150405saaaab.163) where the Si spectrum had a
ZPD time before its corresponding InGaAs spectrum
that *also* just happened to require that the .mav
file go back to the previous 3-hour priors.

I tried to modify the original prior index calculation
to account for this case, but it was just too fragile.
My solution was to use the list of *all* runlog spectra
to match up the spectrum in the .aia file with that
in the .mav file. Now the prior index is calculated by
finding the last spectrum in the .mav that is the current
.aia spectrum or comes before it in the runlog.